Separate DominoUserSession into DominoBasisSession and DominoSetupSession
