+++ src/index.ts
@@ -7,6 +7,7 @@
 /* index have no testable code - no point including them in coverage reports */
 
 import { CredentialType, DominoAccess, DominoRestAccessJSON, RestCredentials } from './DominoAccess';
+import DominoBasisSession from './DominoBasisSession';
 import { DominoRequestOptions, DominoRequestResponse, DominoRestOperation } from './DominoConnector';
 import { DocumentBody, DocumentJSON, DominoBaseDocument, DominoDocumentMeta } from './DominoDocument';
 import {
@@ -47,15 +48,16 @@ import {
 } from './DominoListViewOperations';
 import { AccessLevel, DominoBaseScope, ScopeBody, ScopeJSON } from './DominoScope';
 import { DominoApiMeta, DominoServer } from './DominoServer';
+import DominoSetupSession from './DominoSetupSession';
 import DominoUserSession from './DominoUserSession';
 import {
   ApiNotAvailable,
+  CallbackError,
   EmptyParamError,
   HttpResponseError,
   InvalidParamError,
   MissingBearerError,
   MissingParamError,
-  CallbackError,
   NoResponseBody,
   NotAnArrayError,
   OperationNotAvailable,
@@ -71,6 +73,7 @@ export {
   BulkGetDocumentsOptions,
   BulkGetErrorResponse,
   BulkUpdateDocumentsByQueryRequest,
+  CallbackError,
   CreateDocumentOptions,
   CreateUpdateListResponse,
   CredentialType,
@@ -86,6 +89,7 @@ export {
   DominoBaseListView,
   DominoBaseListViewEntry,
   DominoBaseScope,
+  DominoBasisSession,
   DominoDocumentMeta,
   DominoDocumentOperations,
   DominoRequestOptions,
@@ -93,6 +97,7 @@ export {
   DominoRestAccessJSON,
   DominoRestOperation,
   DominoServer,
+  DominoSetupSession,
   DominoUserSession,
   EmptyParamError,
   GetDocumentOptions,
@@ -110,9 +115,8 @@ export {
   ListViewEntryBody,
   ListViewEntryJSON,
   ListViewEntryOptions,
-  MissingParamError,
   MissingBearerError,
-  CallbackError,
+  MissingParamError,
   NoResponseBody,
   NotAnArrayError,
   OperationNotAvailable,
